+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Browser Scenario: Only positive integers above 0 are now allowed for timeout options
 - helper:serializeTimeout: No longer allows passing anything but an integer above 0
 
+### Fixed
+- Node Scenario: Fix issue where working with timeout the timer would not be cleared correctly after abort, leading to ghost timer
+
 ## [1.7.1] - 2022-08-10
 ### Fixed
 - Fix missing import